requirements.txt
fastapi
tortoise-orm[aiomysql]
uvicorn
pyjwt
passlib[bcrypt]==1.7.4
boto3
pydantic-settings
tensorflow
pydantic
pillow
python-multipart
cryptography
inference-sdk
opencv-python-headless
google-generativeai
reportlab
python-jose[cryptography]==3.3.0
bcrypt==4.0.1